Consolidate audio detectors into single-pass audio.analysis.ffmpeg
Field report (RDP, 2026-07-22): analyses appeared hung; ffmpeg at 0.1%
CPU = I/O-starved. Root cause split: environment I/O (network share /
AV scanning suspected) amplified by our architecture reading the file
3x per audio stream (RISKS R9).

- New audio.analysis.ffmpeg chains ebur128 + silencedetect + astats in
  one filter graph: each audio stream is now decoded exactly once
- Parameter IDs unchanged: presets, rules, reports all untouched
  (detector replaceability, ADR-010)
- Old loudness/silence/clipping detector classes removed; their parsers
  are retained and reused (unit tests unchanged)
- Also: first ruff pass over reports/ package (previously skipped
  because ruff respects .gitignore and reports/ was wrongly ignored)

I/O per WAV job: was ~7 full reads (3 audio passes x streams + probes
+ hash), now ~2 (1 combined pass + hash). 176 tests passing.

+++ src/deepdub_qc/reports/html_renderer.py
@@ -45,7 +45,7 @@ def _fmt_unit(value: object, unit: str | None) -> str:
     return f"{value} {unit}" if unit else f"{value}"
 
 
-def format_expected(expected: object) -> str:
+def format_expected(expected: object) -> str:  # noqa: PLR0911 - one return per payload shape
     """Human-readable rendering of a rule's target value."""
     match expected:
         case ExpectedRange(min=lo, max=hi, unit=unit):
@@ -91,7 +91,7 @@ def format_timecode(seconds: float | None, fps: float = _DEFAULT_TIMECODE_FPS) -
         return "—"
     total = int(seconds)
     frames = int((seconds - total) * fps)  # floor: never point past the event
-    if frames >= int(round(fps)):  # guard against float edge cases
+    if frames >= round(fps):  # guard against float edge cases
         frames = 0
         total += 1
     hours, remainder = divmod(total, 3600)
